🔀 :: (#143) 댓글 삭제 Api

+++ src/main/java/io/github/opgg/music_ward_server/dto/comment/request/CommentRequest.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class CommentRequest {
 
     private String comment;
-    private Integer playlistId;
+    private long playlistId;
 
 }
