+++ src/main/java/io/github/opgg/music_ward_server/dto/comment/request/EditCommentRequest.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class EditCommentRequest {
 
-    private Integer commentId;
+    private long commentId;
     private String comment;
 
 }
